+++ crates/ryft-core/src/errors.rs
@@ -1,10 +1,14 @@
 use thiserror::Error;
 
 use crate::parameters::ParameterError;
+use crate::types::data_types::DataTypeError;
 
 /// Represents errors that can occur in `ryft-core`.
 #[derive(Error, Clone, Debug, Eq, PartialEq, Hash)]
 pub enum Error {
     #[error(transparent)]
     Parameter(#[from] ParameterError),
+
+    #[error(transparent)]
+    DataType(#[from] DataTypeError),
 }