+++ crates/ryft/src/lib.rs
@@ -8,4 +8,6 @@ pub use ryft_core::parameters::{
     Parameter, ParameterError, ParameterPath, ParameterPathSegment, Parameterized, ParameterizedFamily,
     PathPrefixedParameterIterator, Placeholder,
 };
+pub use ryft_core::types::Type;
+pub use ryft_core::types::data_types::{DataType, DataTypeError};
 pub use ryft_macros::{Parameter, Parameterized};