+++ crates/ryft-core/CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 - Added `Parameterized::broadcast_to_parameter_structure`.
-- Introduced more fine-grained error types like `ParameterError`.
+- Added the `Type` trait.
+- Added the `DataType` type.
+- Introduced more fine-grained error types like `ParameterError`, and `DataTypeError`.
 
 ## [0.0.2] - 2026-03-02